Rename package in :internal-testutils
