+++ internal-testutils/src/main/java/com/google/accompanist/internal/test/TestUtils.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2021 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 @file:Suppress("NOTHING_TO_INLINE")
 
-package com.google.accompanist.imageloading.test
+package com.google.accompanist.internal.test
 
 inline fun parameterizedParams(): List<Array<Any>> = emptyList()
