+++ pager/src/sharedTest/kotlin/com/google/accompanist/pager/BaseHorizontalPagerTest.kt
@@ -37,8 +37,8 @@ import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.height
 import androidx.compose.ui.unit.width
-import com.google.accompanist.imageloading.test.combineWithParameters
-import com.google.accompanist.imageloading.test.parameterizedParams
+import com.google.accompanist.internal.test.combineWithParameters
+import com.google.accompanist.internal.test.parameterizedParams
 
 @OptIn(ExperimentalPagerApi::class) // Pager is currently experimental
 abstract class BaseHorizontalPagerTest(
